Fixed data type conversion warnings.

+++ ScaryBugs/ScaryBugs/RWTRateView.m
@@ -130,8 +130,8 @@
 - (void)handleTouchAtLocation:(CGPoint)touchLocation {
     if (!self.editable) return;
     
-    int newRating = 0;
-    for(int i = self.imageViews.count - 1; i >= 0; i--) {
+    NSUInteger newRating = 0;
+    for(NSInteger i = self.imageViews.count - 1; i >= 0; i--) {
         UIImageView *imageView = [self.imageViews objectAtIndex:i];        
         if (touchLocation.x > imageView.frame.origin.x) {
             newRating = i+1;
